# 04-Source-code/docs/api/v1/openapi.yaml
...
openapi: 3.0.3
info:
  title: Alarmsystem-Backend G2 — API v1 (Vereisungserkennung ANR)
  version: 1.0.0
  description: |
    REST-Schnittstelle des G2-Backends (Erfassung & Bewertung von
    Vereisungsbedingungen am Regionalflughafen ANR). **G2 ist hier Server**,
    **G3 (Frontend) ist Client** (GET / SSE-Abo — G3 hostet nichts).

    **Sicherheits-Randbedingung RB-01:** Das System ist reine
    *Entscheidungsunterstuetzung*. Es gibt **keinen** Endpoint, der die Startbahn
    freigibt/sperrt/steuert. Quittieren eines Alarms ist eine reine UI-/Audit-
    Aktion, kein Steuerbefehl.

    **Fail-safe NF-01:** Bei veralteten (stale) oder fehlenden Daten liefert die
    aktuelle Bewertung **nie GRUEN**, sondern `risk_level=unknown` +
    `is_stale=true` mit HTTP `200` — das ist **kein** Fehlerfall. `503` nur, wenn
    G2 selbst nicht lieferfaehig ist (z. B. noch keine Daten / interner Ausfall).

    Quelle: `docs/API_FROZEN_v1.md`. Konsumierter G1-Vertrag: siehe
    `g1-consumed.openapi.yaml`.
servers:
  - url: http://localhost:8000
    description: Lokale Entwicklung (Prototyp)
...
paths:
...
  /v1/assessment/current:
    get:
      tags: [Assessment]
      operationId: getCurrentAssessment
      summary: Aktuelle Vereisungsbewertung
      description: |
        Liefert die zuletzt berechnete Bewertung als **Ampel + Roh-Messwerte**
        in einem Response (flach, kein Envelope — E-36).

        **Fail-safe (NF-01):** Sind die zugrunde liegenden Daten veraltet
        (stale) oder fehlt G1, ist dies **kein** Fehler: Response `200` mit
        `is_stale=true` und `risk_level=unknown` (nie GRUEN). `503` nur, wenn G2
        selbst (noch) nicht lieferfaehig ist.

        **RB-01:** rein beratend — keine Freigabe-/Steueraktion.
      responses:
        "200":
          description: Aktuelle Bewertung (ggf. Fail-safe `unknown` + `is_stale=true`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentCurrent"
              examples:
                normal:
                  summary: Normalfall (Risiko gelb)
                  value:
                    risk_level: yellow
                    driving_factor: dew_point
                    explanation: "ΔT ≤ 1,0 °C an der Oberfläche → Feuchte"
                    surface_temp_c: -0.4
                    dew_point_c: -1.1
                    delta_t: 0.7
                    humidity_pct: 96
                    measured_at: "2026-06-22T14:03:05Z"
                    assessed_at: "2026-06-22T14:03:30Z"
                    is_stale: false
                    sensor_status: ok
                failsafe_stale:
                  summary: Fail-safe — veraltete Daten (NF-01)
                  value:
                    risk_level: unknown
                    driving_factor: null
                    explanation: "Letzter Messwert älter als der Stale-Timeout — keine gültige Bewertung."
                    surface_temp_c: null
                    dew_point_c: null
                    delta_t: null
                    humidity_pct: null
                    measured_at: "2026-06-22T13:55:00Z"
                    assessed_at: "2026-06-22T14:03:30Z"
                    is_stale: true
                    sensor_status: ok
                failsafe_fault:
                  summary: Fail-safe — Sensor defekt trotz aktueller Daten (NF-01, nie GRUEN)
                  value:
                    risk_level: unknown
                    driving_factor: null
                    explanation: "Sensor meldet 'fault' — keine vertrauenswürdige Bewertung möglich."
                    surface_temp_c: null
                    dew_point_c: null
                    delta_t: null
                    humidity_pct: null
                    measured_at: "2026-06-22T14:03:05Z"
                    assessed_at: "2026-06-22T14:03:30Z"
                    is_stale: false
                    sensor_status: fault
        "503":
          description: G2 ist nicht lieferfaehig (z. B. noch keine Daten / interner Ausfall). NICHT fuer Stale.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: SERVICE_UNAVAILABLE
                message: "Noch keine Bewertung verfügbar."
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Statischer API-Key fuer **schreibende** Endpoints (DTB-63, NF-07). Soll-Wert
        serverseitig aus Umgebungsvariable `G2_API_KEY` (kein Secret im Code). Lesende
        Endpoints sind offen. Methode bewusst wechselbar (Basic/JWT spaeter) —
        Naht-Entscheidung beim Architekt.
  schemas:
...
    RiskLevel:
      type: string
      description: |
        Vereisungs-Risikostufe. `unknown` ist der Fail-safe-Zustand (NF-01):
        bei veralteten/fehlenden Daten nie GRUEN.
      enum: [green, yellow, orange, red, unknown]
    SensorStatus:
      type: string
      description: Sensor-/Lieferstatus (aus G1s `GET /current` `status`).
      enum: [ok, fault]
...
    AssessmentCurrent:
      type: object
      description: |
        Aktuelle Vereisungsbewertung. Felder gemaess `Assessment`-Schema
        (src/model/schemas.py): `assessed_at` entspricht dort dem Feld `ts`;
        `measured_at` stammt aus dem verknuepften `Reading` (G1-Snapshot, nicht
        aus `Assessment`). Naht-Felder `is_stale`/`sensor_status` ergaenzen die
        Naht-Sicht. Alle Zeitstempel in UTC (ISO-8601).

        **Fail-safe-Invarianten (NF-01) — verbindlich, von der `assessment/`-Logik
        garantiert (OpenAPI 3.0.3 kann felduebergreifende Regeln nicht hart
        erzwingen; Durchsetzung + Tests liegen im kritischen Pfad):**
        - `risk_level=green` ist NUR zulaessig, wenn `is_stale=false` UND
          `sensor_status=ok`. Bei `is_stale=true` oder `sensor_status=fault`
          gilt `risk_level=unknown` (nie GRUEN).
        - `measured_at` ist auf einer 200-Antwort **immer** gesetzt (auch bei
          `risk_level=unknown`); der Fall „gar keine Daten" wird mit **HTTP 503**
          signalisiert, nicht mit `null`.
        - Bei `risk_level != unknown` sind `surface_temp_c`, `dew_point_c`,
          `delta_t` und `humidity_pct` nicht-null (eine echte
          Bewertung hat immer Messwerte). Genullte Messwerte treten **nur** bei
          `risk_level=unknown` auf — ausgeloest durch `is_stale=true` (veraltet)
          **oder** `sensor_status=fault` (Sensor defekt). `measured_at` kann dabei
          im Fault-Fall mit frischen Daten gesetzt sein (Zeitstempel bekannt, Werte
          aber nicht vertrauenswuerdig); bei reinem Stale traegt es den letzten,
          alten Zeitstempel.
      required: [risk_level, assessed_at, is_stale, sensor_status, measured_at]
      properties:
        risk_level:
          $ref: "#/components/schemas/RiskLevel"
        driving_factor:
          type: string
          nullable: true
          maxLength: 64  # = VARCHAR(64) in migrations/schema.sql (assessment.driving_factor)
          description: Hauptfaktor der Bewertung (z. B. `dew_point`, `surface_temp`).
        explanation:
          type: string
          nullable: true
          maxLength: 512  # = VARCHAR(512) in migrations/schema.sql (assessment.explanation)
          description: Menschenlesbare Begruendung der Stufe.
        surface_temp_c:
          type: number
          nullable: true
          description: Oberflaechentemperatur T_s (°C).
        dew_point_c:
          type: number
          nullable: true
          description: Taupunkt (°C), von G2 berechnet (Magnus).
        delta_t:
          type: number
          nullable: true
          description: ΔT = surface_temp_c − dew_point_c (°C).
        humidity_pct:
          type: number
          nullable: true
          description: Luftfeuchte (%).
        measured_at:
          type: string
          format: date-time
          description: |
            G1-Messzeitpunkt des zugrundeliegenden Readings (UTC). Auf einer 200-Antwort
            immer gesetzt (bei `is_stale=true` = letzter bekannter Zeitstempel). Der Fall
            „gar keine Daten vorhanden" wird mit HTTP 503 signalisiert, nicht mit `null`.
        assessed_at:
          type: string
          format: date-time
          description: G2-Bewertungszeitpunkt (UTC).
        is_stale:
          type: boolean
          description: "`true` = Daten ueberaltert; zusammen mit `risk_level=unknown` = Fail-safe griff (NF-01)."
        sensor_status:
          $ref: "#/components/schemas/SensorStatus"

    # --- Messwert (Historie) ---
...
    Error:
      type: object
      description: Maschinenlesbarer Fehler (Code + Meldung). Keine internen Details/Secrets.
      required: [code, message]
      properties:
        code:
          type: string
          description: Maschinenlesbarer Fehlercode (z. B. BAD_REQUEST, NOT_FOUND, SERVICE_UNAVAILABLE).
        message:
          type: string
          description: Menschenlesbare Fehlermeldung.